fix(keycloak): unblock with workaround for keycloak 23 API changes
Keycloak 23 changed the API to not return sub-groups by default. For
that, a dedicated new API has to be used. Unfortunately that API is not
yet available in the Rust keycloak crate.

For now we figured a semi functional workaround to be using the get
groups API with a search query. This internally resolves to a different
aggregator which does not over optimize the get group API to not return
subgroup data.